+++ database/migrations/2020_01_01_700000_create_blocks_table.php
@@ -14,10 +14,10 @@ class CreateBlocksTable extends Migration
     public function up()
     {
         Schema::create('blocks', function (Blueprint $table) {
-            $table->increments('id');
-            $table->unsignedInteger('page_id');
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('page_id');
             $table->foreign('page_id')->references('id')->on('pages');
-            $table->unsignedInteger('order')->default(0);
+            $table->unsignedBigInteger('order')->default(0);
             $table->text('content')->nullable();
             $table->string('status')->default('draft');
             $table->string('type')->default('paragraph');
