+++ database/migrations/2020_01_01_800000_add_parent_id_to_blocks_table.php
@@ -14,7 +14,7 @@ class AddParentIdToBlocksTable extends Migration
     public function up()
     {
         Schema::table('blocks', function (Blueprint $table) {
-            $table->unsignedInteger('parent_id')->nullable();
+            $table->unsignedBigInteger('parent_id')->nullable();
             $table->foreign('parent_id')->references('id')->on('blocks');
         });
     }
